sim: preserve sim_v2 float64 observation contract

+++ Tsimulation/sim_v2/pushshapes/env.py
@@ -1158,12 +1158,12 @@ class PushShapesEnv(gym.Env):
             else to_image_obs(self._render_world(), self.image_size)
         )
         return {
-            "agent_pos": np.array([pos.x, pos.y], dtype=np.float32),
-            "agent_angle": np.array([self._pusher_body.angle], dtype=np.float32),
+            "agent_pos": np.array([pos.x, pos.y], dtype=np.float64),
+            "agent_angle": np.array([self._pusher_body.angle], dtype=np.float64),
             "object_pose": np.array(
-                [obj.position.x, obj.position.y, obj.angle], dtype=np.float32
+                [obj.position.x, obj.position.y, obj.angle], dtype=np.float64
             ),
-            "goal_pose": np.array(self._goal_pose, dtype=np.float32),
+            "goal_pose": np.array(self._goal_pose, dtype=np.float64),
             "image": image,
         }
